fix(nextjs): improve e2e test by reusing the project
