+++ e2e/next-core/src/utils.ts
