+++ e2e/next-core/tsconfig.json
