+++ e2e/next-core/tsconfig.spec.json
